+++ packages/ui/src/features/core/title-card/title-card.style.ts
@@ -28,6 +28,19 @@ export const useStyles = makeStyles()(theme => ({
       paddingBottom: 0,
     },
   },
+  loadingOverlayContainer: {
+    position: 'relative',
+  },
+  loadingOverlay: {
+    background: 'rgb(255 255 255 / 75%)',
+    position: 'absolute',
+    height: '100%',
+    width: '100%',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    zIndex: 2,
+  },
 }))
 
 export default useStyles